Rework providers to be less generic.
