+++ crates/prelude/src/foundry.rs
@@ -27,7 +27,7 @@ pub use seismic_alloy_network::{
 };
 pub use seismic_alloy_provider::{
     test_utils, SeismicCallExt, SeismicProviderBuilder, SeismicProviderExt, SeismicSignedProvider,
-    SeismicUnsignedProvider, ShieldedCallExt,
+    SeismicUnsignedProvider, ShieldedCallExt, SignedProviderExt,
 };
 pub use seismic_alloy_rpc_types::{
     SeismicCallRequest, SeismicRawTxRequest, SeismicTransactionReceipt as TransactionReceipt,
